zth

Trait FiberEntry

pub trait FiberEntry<Args> {
    type Output: 'static;

    // Required method
    fn call(self, args: Args) -> Self::Output;
}

Trait implemented by callable fiber entry points.

This trait is implemented for FnOnce callables that take either no arguments or tuples up to arity 8.

Required Associated Types

type Output: 'static

Required Methods

fn call(self, args: Args) -> Self::Output

Invokes the entry point with the provided argument tuple.